Run shinyscholar Application

This function runs the shinyscholar application in the user's default web browser.

run_shinyscholar(
  launch.browser = TRUE,
  port = getOption("shiny.port"),
  load_file = NULL
)

Arguments

launch.browser

Whether or not to launch a new browser window.

port

The port for the shiny server to listen on. Defaults to a random available port.

load_file

Path to a saved session file which will be loaded when the app is opened

Value

No return value, called for side effects

Author

Jamie Kass jkass@gradcenter.cuny.edu

Gonzalo E. Pinilla-Buitrago gpinillabuitrago@gradcenter.cuny.edu

Simon E. H. Smart simon.smart@cantab.net

Examples

if(interactive()) {
run_shinyscholar()
}